requirements.txt
fastapi
uvicorn[standard]
PyPDF2
openai
python-multipart
APScheduler
python-dotenv
google-generativeai